Not quite sure what changed here e,e

+++ src/Estruturas/Misc.c
@@ -66,5 +66,5 @@ char* load_file(char const *file_name){
 int get_next_line(char* string) {
 	char line[255];
 	sscanf(string, "%[^\n]\n", line);
-	return strlen (line);
+	return (int)strlen (line);
 }
